software/script/requirements.txt
pyserial
stn2120-pck
